Added tolerances

+++ src/main/java/frc/robot/commands/lift/VerticalShift.java
@@ -4,7 +4,7 @@ import frc.robot.subsystems.Lift;
 
 public class VerticalShift extends Command {
     private int _speed, _targetPhase; //positive means going up, negative means going down
-    private double _liftTolerance = 10; //the motors have to be within 10 encoder ticks in order for the command to stop
+    private int _liftTolerance = 10; //the motors have to be within 10 encoder ticks in order for the command to stop
 
     public VerticalShift(int phase, int speed) {
         requires(Lift.getInstance());
